# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "upload image, then make thumbnail 100x100, returns image id",
        "title": "Image Upload API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/upload/form": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a image from multipart/form-data",
                "parameters": [
                    {
                        "type": "file",
                        "description": "image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/server.Failed"
                        }
                    }
                }
            }
        },
        "/upload/json": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a image from application/json",
                "parameters": [
                    {
                        "description": "image file base64 encoded",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/server.JSONImage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/server.Failed"
                        }
                    }
                }
            }
        },
        "/upload/url": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a image from external url",
                "parameters": [
                    {
                        "type": "string",
                        "format": "url",
                        "description": "image url",
                        "name": "image",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/server.Failed"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.Failed": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "explained error",
                    "type": "string"
                }
            }
        },
        "server.JSONImage": {
            "type": "object",
            "properties": {
                "image": {
                    "description": "base64 encoded content",
                    "type": "string"
                }
            }
        },
        "server.Success": {
            "type": "object",
            "properties": {
                "uuid": {
                    "description": "uploaded image uuid",
                    "type": "string"
                }
            }
        }
    }
}